ai_service_python/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
pydantic
pyjwt
bcrypt
passlib[bcrypt]
python-multipart